Cast core page ids to int

+++ projects/plugins/boost/app/lib/critical-css/source-providers/providers/WP_Core_Provider.php
@@ -27,8 +27,8 @@ class WP_Core_Provider extends Provider {
 	public static function get_critical_source_urls( $context_posts = array() ) {
 		$urls = array();
 
-		$front_page = get_option( 'page_on_front' );
-		$posts_page = get_option( 'page_for_posts' );
+		$front_page = (int) get_option( 'page_on_front' );
+		$posts_page = (int) get_option( 'page_for_posts' );
 
 		if ( ! empty( $front_page ) && empty( $context_posts ) ) {
 			$permalink = get_permalink( $front_page );
